// prisma/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const UserRole = {
  ADMIN: 'ADMIN',
  ACCESS_REQUESTER: 'ACCESS_REQUESTER',
  ACCESS_OPERATOR: 'ACCESS_OPERATOR',
  ACCESS_APPROVER: 'ACCESS_APPROVER',
  ACCESS_MONITOR: 'ACCESS_MONITOR',
  SECURITY_MANAGER: 'SECURITY_MANAGER'
} as const

export type UserRole = (typeof UserRole)[keyof typeof UserRole]


export const PlannedAccessStatus = {
  PENDING_APPROVAL: 'PENDING_APPROVAL',
  PENDING_ACCESS: 'PENDING_ACCESS',
  IN_PROGRESS: 'IN_PROGRESS',
  COMPLETED: 'COMPLETED',
  REJECTED: 'REJECTED',
  CANCELLED: 'CANCELLED',
  EXPIRED: 'EXPIRED',
  RESCHEDULED: 'RESCHEDULED'
} as const

export type PlannedAccessStatus = (typeof PlannedAccessStatus)[keyof typeof PlannedAccessStatus]


export const AuditAction = {
  CREATE: 'CREATE',
  FIELD_UPDATE: 'FIELD_UPDATE',
  STATUS_CHANGE: 'STATUS_CHANGE'
} as const

export type AuditAction = (typeof AuditAction)[keyof typeof AuditAction]